JSRPC: Don't allow stubs to observe GC.
We don't want people to rely on GC to close their stubs for them, because this means a change in GC behavior could break them. We cannot control what V8 does with GC.

To that end, we say that any stub which is leaked will live until the IoContext is destroyed. We'll also log a warning if we detect this.

In order to test this, I also added the `ctx.abort()` method we've talked about adding -- guarded as experimental.