BUGFIX: Prevent underflow during read coalesce
The `dirtyList` is a kj::List with a thin wrapper around its methods.
When we add or remove from the list, we also update the size (total
bytes).

This was fine until we started modifying the size of the elements of the
list. In this case, we put an Entry in the `dirtyList` when we want to
read it from storage, and when the result from storage is ready we
overwrite whatever is in `Entry`. When we would later attempt to remove
the Entry from the `dirtyList`, it would decrement the size past 0 and
cause underflow, which would cause tests to hang.